Factor LogDialog
